Added properly sized icons
Xfce calculates the scaling ratio based on the target size and the
directory name, so if the image file's resolution doesn't match the
directory name the icon will appear too large in the menu.

I've taken the 16 and 32 versions from the Osu website's favicon.
The rest are downscaled/cropped from the 1080 version.
Freedesktop recommends a 48 version and the 256 version is there
for programs that don't check for 1024 versions.

I also ran them trough pngquant to make the filesizes somewhat
smaller with no perceivable difference.